Scaffold the basic notebook CLI
Create a scaffold for the CLI that we will use to publish content.

The CLI will have these basic commands:
- checkDependencies: check binary dependencies and suggest to install
  them. Ideally, we should try to bundle as much things into the
  binary, but there will be certain programs (e.g: pandoc) that will
  be hard to bundle in binary form.
- render: render markdown content and themes files into a folder.
- develop: Run a web server on the rendered content => allow for live
  testing of the site. The server don't include cache-control HTTP
  headers => page should be mostly up-to-date (except for unexpected
  browsers caching).
- setupGithubPages: setup a directory and branch to render
  and publish content for Github Pages. Usually share the same folder
  with where we render local files.
- publishToGithubPages: publish up-to-date content to Github Pages.

We will use klog (Kubernetes logging package) to log things
(this is purely my preference).

spf13/cobra is used to setup the CLI.
Meanwhile, spf13/viper allows for a more structured way to deal with
config values.